[13.0] [PORT] from 11.0
[IMP] avoid reading binary on view load (#20)

* [IMP] avoid reading binary on view load
The first time a user is entering a view, odoo is fetching the data of all the binaries that are on reports added to that view.
With this hack we avoid that read and data transfer

* [IMP] avoid reading binary on view load
The first time a user is entering a view, odoo is fetching the data of all the binaries that are on reports added to that view.
With this hack we avoid that read and data transfer